Initial Commit
The current state of the CLI program prompts a user for a subreddit, displays
the top 10 submissions for that subreddit and then asks if they'd like to view
any of those submissions. If you choose a submission, the link is opened in w3m
and the user is directed back to the application after they exit the web
browser.

The user will keep being prompted to choose another submission to view until
they enter a non-numeric input or no input at all.